Version (1, 93, 5)

+++ Blender 2.79/PES_Face_Hair_Modifier.py
@@ -11,7 +11,7 @@ icons_collections = {}
 bl_info = {
 	"name": "PES Face/Hair Modifier",
 	"author": "the4chancup - MjTs-140914",
-	"version": (1, 93, 4),
+	"version": (1, 93, 5),
 	"blender": (2, 79, 0),
 	"api": 35853,
 	"location": "Under Scene Tab",
@@ -22,15 +22,15 @@ bl_info = {
 	"category": "System"
 }
 
-TEMPPATH = str()
-TEMPPATH = os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)), '..'))
-GZSPATH = '"%s\\addons\\Data\\Gzs\\GzsTool.exe"' % TEMPPATH 
-FtexTools ='"%s\\addons\\Data\\Gzs\\FtexTools.exe"' % TEMPPATH 
-texconvTools = '"%s\\addons\\Data\\Gzs\\texconv.exe"' % TEMPPATH 
-ini_sett = '%s\\addons\\Data\\Gzs\\Settings.ini' % TEMPPATH
-xml_sett = '%s\\addons\\Data\\Gzs\\PesFoxShader.xml' % TEMPPATH
-icons_dir = '%s\\addons\\Data\\Gzs\\icons' % TEMPPATH
-base_file_blend = '%s\\addons\\Data\\Gzs\\base_file.blend' % TEMPPATH
+AddonsPath = str()
+AddonsPath = os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)), '..'))
+GZSPATH = '"%s\\addons\\Data\\Gzs\\GzsTool.exe"' % AddonsPath 
+FtexTools ='"%s\\addons\\Data\\Gzs\\FtexTools.exe"' % AddonsPath 
+texconvTools = '"%s\\addons\\Data\\Gzs\\texconv.exe"' % AddonsPath 
+ini_sett = '%s\\addons\\Data\\Gzs\\Settings.ini' % AddonsPath
+xml_sett = '%s\\addons\\Data\\Gzs\\PesFoxShader.xml' % AddonsPath
+icons_dir = '%s\\addons\\Data\\Gzs\\icons' % AddonsPath
+base_file_blend = '%s\\addons\\Data\\Gzs\\base_file.blend' % AddonsPath
 
 pes_diff_bin_data, IDoldname = [] , []
 eyeL_origin, eyeR_origin, mouth_origin = [0.02807705,0.1448301-0.00362,1.69668636147], [0.02754705,0.1446231-0.00362,1.69682636147], [-0.000365,0.1479722561,1.63275]
@@ -563,12 +563,12 @@ class FMDL_Scene_Open_Image(bpy.types.Operator, bpy_extras.io_utils.ImportHelper
 		image.source = 'FILE'
 		image.filepath=filePath
 
-		mesh=bpy.context.scene.objects.active.data
+		mesh=context.scene.objects.active.data
 		uvdata = str()
 
-		texName=bpy.context.active_object.active_material.active_texture.name
+		texName=context.active_object.active_material.active_texture.name
 		texture_role = bpy.data.textures[texName].fmdl_texture_role
-		idx = bpy.context.object.active_material.active_texture_index
+		idx = context.object.active_material.active_texture_index
 		
 		if texture_role == str():
 			self.report({"WARNING"}, "Need set shader first!!")
@@ -581,8 +581,8 @@ class FMDL_Scene_Open_Image(bpy.types.Operator, bpy_extras.io_utils.ImportHelper
 		if mesh.uv_textures.active_index == idx:
 			uvdata = mesh.uv_textures[idx].data
 
-		bpy.context.active_object.active_material.active_texture.name = "[%s] %s" % (texture_role, fileName)
-		texname=bpy.context.active_object.active_material.active_texture.name
+		context.active_object.active_material.active_texture.name = "[%s] %s" % (texture_role, fileName)
+		texname=context.active_object.active_material.active_texture.name
 		bpy.data.textures[texname].fmdl_texture_filename = fileName
 
 		for face in uvdata:
@@ -749,13 +749,14 @@ class FMDL_Scene_Panel_FMDL_Export_Settings(bpy.types.Menu):
 	bl_label = "Export settings"
 
 	def draw(self, context):
-		self.layout.prop(context.active_object, 'fmdl_export_extensions_enabled')
-		row = self.layout.row()
-		row.prop(context.active_object, 'fmdl_export_loop_preservation')
-		row.enabled = context.active_object.fmdl_export_extensions_enabled
-		row = self.layout.row()
-		row.prop(context.active_object, 'fmdl_export_mesh_splitting')
-		row.enabled = context.active_object.fmdl_export_extensions_enabled
+		if context.active_object is not None:
+			self.layout.prop(context.active_object, 'fmdl_export_extensions_enabled')
+			row = self.layout.row()
+			row.prop(context.active_object, 'fmdl_export_loop_preservation')
+			row.enabled = context.active_object.fmdl_export_extensions_enabled
+			row = self.layout.row()
+			row.prop(context.active_object, 'fmdl_export_mesh_splitting')
+			row.enabled = context.active_object.fmdl_export_extensions_enabled
 
 
 class FMDL_Object_BoundingBox_Create(bpy.types.Operator):
@@ -1716,8 +1717,6 @@ class FMDL_Reload_Image(bpy.types.Operator):
 				bpy.ops.object.mode_set(mode='EDIT')
 				image.reload()
 				bpy.ops.object.mode_set(mode='OBJECT')
-			else:
-				bpy.data.images.remove(image)
 		self.report({"INFO"}, "All image texture reloaded!")
 		return {'FINISHED'}
 	pass
@@ -1766,7 +1765,7 @@ class Fmdl_UIPanel(bpy.types.Panel):
 		this_icon = icons_collections["custom_icons"]["icon_0"].icon_id
 		row.label(text="Made by: MjTs-140914 / the4chancup", icon_value=this_icon)
 		row = box.row()
-		box.label(text="This Tool Works with Only Blender v2.79 (v1.93.4b)", icon="BLENDER")
+		box.label(text="This Tool Works with Only Blender v2.79 (v1.93.5b)", icon="BLENDER")
 		row = box.row()
 		row.operator("primary.operator", text="Start New Scene").face_opname = "newscene"
 		row = box.row()
@@ -2129,6 +2128,15 @@ class Tool_Main_Operator(bpy.types.Operator):
 			return {'FINISHED'}
 
 		if self.face_opname == "clr_file":
+			if context.scene.face_cnf == False:
+				self.report({"WARNING"}, "Face ID has change you need to export!")
+				return {'CANCELLED'}
+			if context.scene.hair_cnf == False:
+				self.report({"WARNING"}, "Hair ID has change you need to export!")
+				return {'CANCELLED'}
+			if context.scene.fpk_cnf == False:
+				self.report({"WARNING"}, "ID has change you need to create .fpk!")
+				return {'CANCELLED'}
 			pes_diff_bin_data.clear()
 			if os.path.isfile(packfpk):
 				os.remove(packfpk)
